Refactor dashboard header into reusable Header component
Extracted the dashboard header markup and logic into a new reusable Header component in components/Header.tsx. Updated DashboardPage to use the new Header component for improved code reuse and maintainability.
